+++ flutter/lib/src/integrations/on_error_integration.dart
@@ -13,8 +13,9 @@ typedef ErrorCallback = bool Function(Object exception, StackTrace stackTrace);
 ///
 /// Remarks:
 /// - Only usable on Flutter >= 3.3.0.
+/// - Does not work on Flutter Web
 ///
-/// This can be used instead of the [RunZonedGuardedIntegration]. Removing the
+/// This is used instead of [RunZonedGuardedIntegration]. Not using the
 /// [RunZonedGuardedIntegration] results in a minimal improved startup time,
 /// since creating [Zone]s is not cheap.
 class OnErrorIntegration implements Integration<SentryFlutterOptions> {
@@ -59,6 +60,12 @@ class OnErrorIntegration implements Integration<SentryFlutterOptions> {
         // ignore: invalid_use_of_internal_member
         timestamp: options.clock(),
       );
+
+      // marks the span status if none to `internal_error` in case there's an
+      // unhandled error
+      hub.configureScope((scope) => {
+            scope.span?.status ??= const SpanStatus.internalError(),
+          });
 
       // unawaited future
       hub.captureEvent(event, stackTrace: stackTrace);